pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "building-energy-management"
version = "1.0.0"
description = "Edge AI Building Energy Management System for Research and Education"
authors = [
    {name = "Edge AI Research Team", email = "research@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core ML and data science
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    
    # Deep learning frameworks
    "tensorflow>=2.13.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0",
    
    # Edge deployment and optimization
    "tensorflow-model-optimization>=0.7.0",
    "tf2onnx>=1.15.0",
    "openvino>=2023.0.0",
    
    # IoT and messaging
    "paho-mqtt>=1.6.0",
    "kafka-python>=2.0.0",
    "websockets>=11.0.0",
    
    # Signal processing and computer vision
    "opencv-python>=4.8.0",
    "scikit-image>=0.21.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    
    # MLOps and monitoring
    "mlflow>=2.5.0",
    "wandb>=0.15.0",
    
    # Web interface
    "streamlit>=1.25.0",
    "gradio>=3.40.0",
    
    # Utilities
    "psutil>=5.9.0",
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.3.0",
    "mypy>=1.5.0",
]

edge = [
    "tensorrt>=8.6.0",
    "coremltools>=7.0.0",
    "tvm>=0.14.0",
]

federated = [
    "flower>=1.5.0",
    "fedml>=0.8.0",
    "grpcio>=1.56.0",
]

industrial = [
    "opcua>=0.98.0",
    "pymodbus>=3.4.0",
]

[project.urls]
Homepage = "https://github.com/edge-ai-research/building-energy-management"
Repository = "https://github.com/edge-ai-research/building-energy-management.git"
Documentation = "https://github.com/edge-ai-research/building-energy-management#readme"
Issues = "https://github.com/edge-ai-research/building-energy-management/issues"

[project.scripts]
bems-train = "src.scripts.train:main"
bems-evaluate = "src.scripts.evaluate:main"
bems-export = "src.scripts.export:main"
bems-demo = "src.scripts.demo:main"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "edge: marks tests as edge deployment tests",
]
